# docs/api-snapshot/openapi.json
{
  "components": {
    "schemas": {
      "CreateEdgeBody": {
        "properties": {
          "dst": {
            "title": "Dst",
            "type": "string"
          },
          "edge_type": {
            "title": "Edge Type",
            "type": "string"
          },
          "facet_binding": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Facet Binding"
          },
          "mode": {
            "default": "track",
            "title": "Mode",
            "type": "string"
          },
          "pinned_commit": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Pinned Commit"
          },
          "provenance": {
            "title": "Provenance",
            "type": "string"
          },
          "src": {
            "title": "Src",
            "type": "string"
          }
        },
        "required": [
          "src",
          "dst",
          "edge_type",
          "provenance"
        ],
        "title": "CreateEdgeBody",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "title": "tm-daemon API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/v1/edges": {
      "post": {
        "operationId": "create_edge_v1_edges_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateEdgeBody"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Create Edge V1 Edges Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Edge",
        "tags": [
          "edges"
        ]
      }
    },
    "/v1/edges/{edge_id}": {
      "delete": {
        "operationId": "delete_edge_v1_edges__edge_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "edge_id",
            "required": true,
            "schema": {
              "title": "Edge Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Delete Edge V1 Edges  Edge Id  Delete",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Delete Edge",
        "tags": [
          "edges"
        ]
      }
    },
...
  }
}